+++ vers/changelog.md
@@ -6,7 +6,8 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* Fix: authHeaderEvaluator for WCF now passed methodInfo also for Methods from inherited contracts (was null before)
+* AuthHeaderEvaluator can now return a failedReason byref